add Seeed_Arduino_CAN as a submodule
This was done since it does not appear to be in the arduino library
manager, and thus would need to be cloned anyway. I still have not yet
decided if this is how all arduino libraries will be installed, or if
they should be installed using the arduino library manager.