remove can library submodule
I found out that the arduino build system is bad and incompatible with
headers and sources in different directories. It appears that
unfortunately, the best system is to download both the dependency
library and this library to arduino/libraries directory separately
instead of just this one containing the other. Maybe one day we can find
a better solution to this.