make arduino library
Because the arduino build system seems to strongly enforce its structure
of libraries being in $ARDUINOHOME/libraries, this is being turned into
an arduino library. It should include the driver libraries as
dependencies, but this could not yet be figured out. For now, they will
need to be installed separately as instructed in the README.