implement sendMessage
Send message is now implemented. I still am not a fan of the type
casting, but the thought was that I would prefer to use standard types
of void*/size_t, etc. However, perhaps it would be better to use arduino
types, as we are using an arduino here. I still think I'm wanting to
make a can message struct, which would replace the void*, but this
definitely depends on what information we're sending and receiving. Just
not enough information yet.